+++ NEWS.md
@@ -1,3 +1,10 @@
+# RcppColors 0.3.0
+
+- New color mapping function `colorMap3`.
+
+- The three color mapping functions can be executed using parallel computations.
+
+
 # RcppColors 0.2.0
 
 New function `permuteRGB`, to permute the color channels of a color.